Refactor (#18)
* moved data-related functions around
* add options and paths modules
* add --rad-data command-line option
* using the same IR model numbers as in the notebook
* remove unnecessary import
* handle the case where there is not monitoring windows
* bugfix
* add some stdout to low_kalman_mon for a typical error when debugging
* in kalman_watch_monitor_win_perigee, set default output name using start/stop and allow saving to png
* remove spurious f
* changes to figure style
* fix in kalman_watch_data cache: create directory including parents if it does not exist
* add to gitignore
* use a mon_win_kalman_drops_-45d_-1d.html as default in-file in low_kalman_mon.py, and fix docstring in monitor_win_perigee.py
* use default in/out filename in task schedule
* use in-file from command line options in low_kalman_mon.py
* increase height of perigee kalman plot
* renamed options.OPTIONS -> config.conf
* rename rootname in kalman_watch.config.ConfigNamespace
* fix some pylance warnings
* stop using astropy config
* fixup
* removed duplicated code related to get_kalman_drops_npnt:
- added n_cache to config
- renamed EventPerigee.get_kalman_drops_npnt as a property called kalman_drops_npnt,
- renamed get_kalman_drops_nman as a property called kalman_drops_nman,
- fixed wrong property caching for predicted_kalman_drops
- remove duplicated code in _get_kalman_drops_nman
- refactor get_kalman_drops_nman so it has the same signature as get_kalman_drops_npnt
- removed _get_binned_drops_from_npnt and use get_binned_drops_from_npnt instead,
* set argparse defaults from config
* ruff
* rename radiation data file
